#Conflicts:
#	DormWars/Models/Event.swift

+++ DormWars/Models/EventTeam.swift
@@ -5,14 +5,14 @@
 //  Created by Drew Mendelow on 9/27/25.
 //
 class EventTeam: Codable {
-    var eventTeamId: Int
-    var eventId: Int
+    var eventTeamId: Int64
+    var eventId: Int64
     var team: Team
     var isTournamentTeam: Bool
     var tournament: Tournament?
     var active: Bool
     
-    init(eventTeamId: Int, eventId: Int, team: Team, isTournamentTeam: Bool, tournament: Tournament? = nil, active: Bool) {
+    init(eventTeamId: Int64, eventId: Int64, team: Team, isTournamentTeam: Bool, tournament: Tournament? = nil, active: Bool) {
         self.eventTeamId = eventTeamId
         self.eventId = eventId
         self.team = team
